fix: chat messages ordering between iOS and Android Root cause: Firestore sorts by data type before by value (Number < Timestamp). If iOS stores Int64 and Android/gitlive stores Long as Timestamp, all messages from one platform appear before the other regardless of actual send time. Fix: - Both platforms now store timestamp as Double (Firestore Number type) - iOS ChatView: Int64 → Double for consistent Firestore type - Android ChatRepositoryImpl: Long.toDouble() for same reason - iOS listener: added client-side .sorted { } as defensive fallback - iOS reader: prioritizes Double cast before Int64 (matches new write format)

+++ shared/src/commonMain/kotlin/com/afilaxy/data/repository/ChatRepositoryImpl.kt
@@ -18,7 +18,10 @@ class ChatRepositoryImpl(
                 "senderId" to message.senderId,
                 "senderName" to message.senderName,
                 "message" to message.message,
-                "timestamp" to message.timestamp,
+                // Double garante tipo Firestore Number — mesmo tipo que o iOS grava.
+                // Kotlin Long pode ser serializado pelo gitlive SDK como Timestamp em alguns
+                // contextos, causando agrupamento por plataforma na query orderBy("timestamp").
+                "timestamp" to message.timestamp.toDouble(),
                 "isFromHelper" to message.isFromHelper
             )
